effect of being banned
